Modified and adjusted model classes. HTTP requests assembled and verified for occasional throw that may occur.
As I'm still learning Flutter, my mind still works based on Activitys, then the files within the folder will be Activitys all views and project pages.

Added the GridView package to update the grid according to the data. I'm wondering if structurally is better to leave it in Activitys folder or in a separate folder.

Added Toast dependency for testing and code verification purposes only.