A new screen was added by pulling the Title, CoverUrl, GridView ID information from the previous screen.
Removed Service Class.

Class of requests for the API has been modified to add a standard url if the API does not have an image for the game.

Image bug not being displayed and an error called: "Invalid argument (s): Unsupported scheme 'file' in URI" still persists.

Drawer menu added with the platform API resquest. (Still without a Title Drawer, but this can be added later)